+++ code/packages/typescript/spice-netlist-parser/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Parse SPICE `.temp <celsius> [celsius ...]` operating-temperature cards.
 - Parse MOS Level-1 capacitance parameters with `.model ... NMOS|PMOS(... CGSO=<c>
   CGDO=<c> CGBO=<c> CBS=<c> CBD=<c>)`.
 - Parse diode model-card emission coefficients with `.model ... D(... N=<n>)`